Fix: エラー修正?

+++ pages/api/account/delete.ts
@@ -1,10 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import {
-    DeleteObjectCommand ,
-    S3Client
-  } from '@aws-sdk/client-s3'
-import jwt from "jsonwebtoken"
-import { createClient } from '@supabase/supabase-js';
+import { DeleteObjectCommand, S3Client } from "@aws-sdk/client-s3";
+import jwt from "jsonwebtoken";
+import { createClient } from "@supabase/supabase-js";
 
 const supabaseAdmin = createClient(
   process.env.NEXT_PUBLIC_SUPABASE_URL as string,
@@ -12,47 +9,35 @@ const supabaseAdmin = createClient(
 );
 
 const r2 = new S3Client({
-    region: 'auto',
-    endpoint: `https://${process.env.CLOUDFLARE_R2_ACCOUNT_ID}.r2.cloudflarestorage.com`,
-    credentials: {
-        accessKeyId: process.env.CLOUDFLARE_R2_ACCESS_KEY as string,
-        secretAccessKey: process.env.CLOUDFLARE_R2_SECRET_KEY as string,
-    },
+  region: "auto",
+  endpoint: `https://${process.env.CLOUDFLARE_R2_ACCOUNT_ID}.r2.cloudflarestorage.com`,
+  credentials: {
+    accessKeyId: process.env.CLOUDFLARE_R2_ACCESS_KEY as string,
+    secretAccessKey: process.env.CLOUDFLARE_R2_SECRET_KEY as string,
+  },
 });
 
 const Delete = async (req: NextApiRequest, res: NextApiResponse) => {
-    if (req.method !== "POST") {
-        return res.status(405).json({ message: "Method not allowed" });
-    }
+  if (req.method !== "POST") {
+    return res.status(405).json({ message: "Method not allowed" });
+  }
 
-    const { token } = req.body;
+  const { token } = req.body;
 
-    try {
-        var decoded = jwt.decode(token)
+  try {
+    var decoded = jwt.decode(token);
 
-        await supabaseAdmin
-            .from("images")
-            .delete()
-            .match({ user_id: decoded.sub })
+    await supabaseAdmin.from("images").delete().match({ user_id: decoded.sub });
 
-        await supabaseAdmin
-            .from("likes")
-            .delete()
-            .match({ user_id: decoded.sub })
+    await supabaseAdmin.from("likes").delete().match({ user_id: decoded.sub });
 
-        await supabaseAdmin
-            .from("profiles")
-            .delete()
-            .match({ id: decoded.sub })
+    await supabaseAdmin.from("profiles").delete().match({ id: decoded.sub });
 
-        await supabaseAdmin.auth.api.deleteUser(
-            decoded.sub
-        )
-        
-    }catch (err) {
-        return res.status(403).json({"error": err});
-    }
-    return res.status(200);
+    await supabaseAdmin.auth.api.deleteUser(decoded.sub);
+  } catch (err) {
+    return res.status(403).json({ error: err });
+  }
+  return res.status(200);
 };
 
 export default Delete;
